requirements.txt
asgiref==3.5.2
Django==3.2.11
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
djangorestframework==3.13.1
django_environ==0.8.1
psycopg2-binary==2.9.3
pillow==9.1.1
twilio==7.9.1
django_rest_passwordreset==1.2.1
djangorestframework-simplejwt==5.2.0
django-rest-auth==0.9.5
django-allauth==0.50.0
django-cors-headers==3.10.1
whitenoise==6.1.0
cloudinary==1.29.0
django-cloudinary-storage==0.3.0
gunicorn==20.1.0
django-heroku==0.3.1

